requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
httpx>=0.24.0
pyyaml>=6.0
python-dotenv>=1.0.0